Fix the issue in update cart details instantlly after Add cart and Delete item from cart.

+++ src/components/CartButton.js
@@ -36,14 +36,17 @@ useEffect(() => {
 
 //Detele the cart from Add cart popup
 const delete_cart_item = (cart_item_id) => {    
-  dispatch(removeItemFromCart(selector.addcart.cartId, cart_item_id));
+  var result = confirm("Are you sure you would like to remove this item from the shopping cart?");
+  if (result) {
+    dispatch(removeItemFromCart(selector.addcart.cartId, cart_item_id)); 
+  }
 }
 
   return (            
             <div>
                 <div className="row" id="Popover1">                    
-                  <div className="text-left p-1"><i className="fa fa-shopping-cart fa-2x text-center pt-2" aria-hidden="true" id="icon" data-toggle="modal" data-target="#loginModal"></i> </div>
-  <div className="text-right p-1 pt-3">{selector.addcart.cartData.cart ? <div>{selector.addcart.cartData.cart.items.length == 0 ? null : <div className="text-light bg-danger pl-2 pt-1 pr-2 pb-1">{selector.addcart.cartData.cart.items.length}</div>}</div> : null }</div>
+                    <div className="text-left p-1"><i className="fa fa-shopping-cart fa-2x text-center pt-2" aria-hidden="true" id="icon"></i> </div>
+                    <div className="text-right p-1 pt-3">{selector.addcart.cartData.cart ? <div>{selector.addcart.cartData.cart.items.length == 0 ? null : <div className="text-light bg-danger pl-2 pt-1 pr-2 pb-1">{selector.addcart.cartData.cart.items.length}</div>}</div> : null }</div>
                 </div>
                 <Popover placement="bottom" isOpen={popoverOpen} target="Popover1" toggle={toggle}>                
                 <PopoverBody>
@@ -57,8 +60,7 @@ const delete_cart_item = (cart_item_id) => {
                                 {selector.addcart.cartData.cart.items.length == 0 ?
                                 <div className="text-center text-secondary pt-3 pb-3">You have no items in your shopping cart.  </div>
                                 :
-                                <div className="pt-3 p-2">
-                                    {/*<h6>Guest CartID:  {selector.addcart.cartId}</h6>*/}
+                                <div className="pt-3 p-2">                        
                                     
                                     <div className="clearfix p-1">
                                         <div className="float-left">{selector.addcart.cartData.cart.items.length == 1 ? <div>1 Item in Cart</div> : <div>{selector.addcart.cartData.cart.items.length} Items in cart</div>}</div>
@@ -74,14 +76,14 @@ const delete_cart_item = (cart_item_id) => {
                                       <div id={cartItem.id}>
                                         <div className="row">
                                           <div className="col-4">
-                                            <div style={{width: '100px', height: '100px'}}>                                            
-                                              <img className="mx-auto p-1" src={cartItem.product.small_image.url} alt="new" style={{width: '100%', height: '100%', objectFit: 'contain'}} />
+                                            <div style={{width: '75px', height: '75px'}}>                                            
+                                              <img className="mx-auto" src={cartItem.product.small_image.url} alt="new" style={{width: '100%', height: '100%', objectFit: 'contain'}} />
                                             </div>
                                           </div>
                                           <div className="col-8">
-                                            <p>{cartItem.product.name}</p>
-                                            <p><b>{getSymbolFromCurrency(cartItem.product.price.regularPrice.amount.currency)}&nbsp;{cartItem.product.price.regularPrice.amount.value}</b></p>
-                                            <p>
+                                            <p className="m-0">{cartItem.product.name}</p>
+                                            <p className="m-0"><b>{getSymbolFromCurrency(cartItem.product.price.regularPrice.amount.currency)}&nbsp;{cartItem.product.price.regularPrice.amount.value}</b></p>
+                                            <p className="m-0">
                                               Qty: <span className="pl-2"><input type="text" value={cartItem.quantity} className="p-1 text-center" style={{width: '50px', height: '30px'}}/></span>
                                               <div className="ml-auto text-secondary text-right">
                                                 <i className="fa fa-edit fa-1x pl-2 pr-2" aria-hidden="true" id="icon"></i>
@@ -92,11 +94,10 @@ const delete_cart_item = (cart_item_id) => {
 
                                       </div>                                      
                                       <hr className="text-secondary p-1"></hr>                                      
-                                      <p className="text-center text-primary p-1" onClick={toggleClose}><Link to={"/checkout/cart"}>View and Edit Cart</Link></p>
+                                      
                                       </div>
                                     ))}
-
-                                    
+                                    <p className="text-center text-primary m-0" onClick={toggleClose}><Link to={"/checkout/cart"}>View and Edit Cart</Link></p>                                    
                                 </div>
                                 }
                             </div>
